+++ crates/arpa-node/src/listener/ready_to_handle_randomness_task.rs
@@ -1,12 +1,13 @@
 use super::Listener;
 use crate::{
-    context::{BLSTasksHandler, BlockInfoHandler, ChainIdentityHandlerType, GroupInfoHandler},
+    context::ChainIdentityHandlerType,
     error::NodeResult,
     event::ready_to_handle_randomness_task::ReadyToHandleRandomnessTask,
     queue::{event_queue::EventQueue, EventPublisher},
 };
 use arpa_contract_client::adapter::AdapterViews;
 use arpa_core::RandomnessTask;
+use arpa_dal::{BLSTasksHandler, BlockInfoHandler, GroupInfoHandler};
 use async_trait::async_trait;
 use ethers::{providers::Middleware, types::Address};
 use log::info;
