+++ crates/arpa-node/src/subscriber/randomness_signature_aggregation.rs
@@ -1,7 +1,7 @@
 use super::{DebuggableEvent, DebuggableSubscriber, Subscriber};
 use crate::{
     algorithm::bls::{BLSCore, SimpleBLSCore},
-    context::{BlockInfoHandler, ChainIdentityHandlerType, SignatureResultCacheHandler},
+    context::ChainIdentityHandlerType,
     error::{NodeError, NodeResult},
     event::{ready_to_fulfill_randomness_task::ReadyToFulfillRandomnessTask, types::Topic},
     queue::{event_queue::EventQueue, EventSubscriber},
@@ -13,6 +13,7 @@ use arpa_core::{
     DEFAULT_MAX_RANDOMNESS_FULFILLMENT_ATTEMPTS,
 };
 use arpa_dal::{cache::RandomnessResultCache, BLSResultCacheState};
+use arpa_dal::{BlockInfoHandler, SignatureResultCacheHandler};
 use async_trait::async_trait;
 use ethers::types::Address;
 use log::{debug, error, info};
