--- crates/dal/sqlite/src/core/mod.rs
@@ -1,7 +1,0 @@
-mod mutation;
-mod query;
-
-pub use mutation::*;
-pub use query::*;
-
-pub use sea_orm;
