+++ crates/dal/sqlite/src/result/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod main;
+mod op;
+
+pub use base::BaseSignatureResultDBClient;
+pub use main::SignatureResultDBClient;
+pub use op::OPSignatureResultDBClient;
