+++ crates/dal/sqlite/src/task/mod.rs
@@ -1,0 +1,10 @@
+mod base;
+mod main;
+mod op;
+
+pub use base::BaseBLSTasksDBClient;
+pub use base::BaseRandomnessTaskQuery;
+pub use main::BLSTasksDBClient;
+pub use main::RandomnessTaskQuery;
+pub use op::OPBLSTasksDBClient;
+pub use op::OPRandomnessTaskQuery;
